Move overlay into src/node/overlay.ts
